requirements.txt
pdfminer.six
sentence-transformers
groq
python-dotenv
streamlit